Added Check for Internet Connection
Added a check for internet connection before trying to search for remote versions.  If not internet connection was found the script would error and crash out.  Going forward, If no internet connection is found the script will not proceed further in the script and exit with a status comment for logging and status code 2.
Moved 'filterArray' to the top for easier edits in the future.